pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "creativeiq"
version = "0.1.0"
description = "CreativeIQ — creative intelligence platform for ad optimization"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Bhargavi", email = "bhargavi@weeklyweights.example" }]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.14.0",
    "pydantic[email]>=2.10.0",
    "email-validator>=2.0.0",
    "pydantic-settings>=2.6.0",
    "asyncpg>=0.30.0",
    "redis>=5.2.0",
    "celery>=5.4.0",
    "structlog>=24.4.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.1.0",
    "python-multipart>=0.0.12",
    "uuid-utils>=0.9.0",
    "pgvector>=0.3.6",
    "boto3>=1.35.0",
    "aioboto3>=13.0.0",
]

[project.optional-dependencies]
google = ["google-cloud-storage>=2.18.0", "google-generativeai>=0.8.0"]
analytics = [
    "snowflake-connector-python[pandas]>=3.12.3",
    "pandas>=2.2.3",
    "dbt-core>=1.10.0",
    "dbt-snowflake>=1.10.0",
    "apache-airflow>=2.10.3",
]
decomposition = [
    "ffmpeg-python>=0.2.0",
    "opencv-python-headless>=4.10.0",
    "librosa>=0.10.2",
    "soundfile>=0.12.1",
    "Pillow>=10.4.0",
    "openai-whisper>=20231117",
    "prometheus_client>=0.21.0",
    "psycopg2-binary>=2.9.10",
]
profile_engine = [
    "scipy>=1.14.1",
    "scikit-learn>=1.5.2",
    "snowflake-connector-python[pandas]>=3.12.3",
]
chat = [
    "langgraph>=0.2.50",
    "langchain-core>=0.3.20",
]
generation = [
    "langgraph>=0.2.50",
    "langchain-core>=0.3.20",
    "python-docx>=1.1.2",
    "pypdf>=5.1.0",
    "reportlab>=4.2.5",
    "prometheus_client>=0.21.0",
    "snowflake-connector-python[pandas]>=3.12.3",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "factory-boy>=3.3.0",
    "ruff>=0.8.0",
    "Pillow>=10.4.0",
    "prometheus_client>=0.21.0",
    "scipy>=1.14.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["shared*", "services*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
